fix(trusty-memory): block AWS access key IDs in content-gate secret scanner (refs #1481)
FN-1 (HIGH): AWS access key IDs (`AKIA…` long-term, `ASIA…` STS temp creds)
are all-uppercase base32, so they fail the `has_lower && has_upper && has_digit`
mixed-case heuristic in `looks_like_secret` and slipped through the content gate
even though the doc comments claimed `AKIA…` keys were caught. This was a real
secret-leak path into palace storage that contradicted the code's own docs.

- Add `akia`/`asia` to SECRET_PREFIXES (compared lowercased, matching the
  existing `sk-`/`ghp_` convention) so AWS key IDs are flagged by prefix.
- Correct the now-accurate doc comments on SECRET_PREFIXES and looks_like_secret
  so they describe what the code actually matches (the mixed-case fallback does
  NOT catch all-uppercase AWS keys; the prefix list does).
- Add `aws_access_key_ids_are_blocked` (AKIA + ASIA reject as PotentialSecret)
  and `git_sha_still_allowed_after_aws_fix` (40-hex SHA over-block guard).
- Extract the test module into `filter_tests.rs` via `#[path]` to keep the
  production file under the 500-SLOC cap.

Verified: cargo test (24/24), clippy -D warnings clean, fmt --check clean,
check_line_cap.sh exit 0.
